formatting

+++ src/main/java/com/gmail/goosius/siegewar/settings/SiegeWarSettings.java
@@ -502,15 +502,15 @@ public class SiegeWarSettings {
 		return Settings.getInt(ConfigNodes.SPECIAL_VICTORY_EFFECTS_DECISIVE_VICTORY_THRESHOLD);
 	}
 
-	public static double getSpecialVictoryEffectsWarchestReductionPercentageOnCloseVictory() {
-		return Settings.getDouble(ConfigNodes.SPECIAL_VICTORY_EFFECTS_WARCHEST_REDUCTION_PERCENTAGE_ON_CLOSE_VICTORY);
+	public static int getSpecialVictoryEffectsWarchestReductionPercentageOnCloseVictory() {
+		return Settings.getInt(ConfigNodes.SPECIAL_VICTORY_EFFECTS_WARCHEST_REDUCTION_PERCENTAGE_ON_CLOSE_VICTORY);
 	}
 
-	public static double getSpecialVictoryEffectsPlunderReductionPercentageOnCloseVictory() {
-		return Settings.getDouble(ConfigNodes.SPECIAL_VICTORY_EFFECTS_PLUNDER_REDUCTION_PERCENTAGE_ON_CLOSE_VICTORY);
+	public static int getSpecialVictoryEffectsPlunderReductionPercentageOnCloseVictory() {
+		return Settings.getInt(ConfigNodes.SPECIAL_VICTORY_EFFECTS_PLUNDER_REDUCTION_PERCENTAGE_ON_CLOSE_VICTORY);
 	}
 
-	public static double getSpecialVictoryWeaknessOnRevoltSiegeDecisiveDefenderVictory() {
+	public static int getSpecialVictoryWeaknessOnRevoltSiegeDecisiveDefenderVictory() {
 		return Settings.getInt(ConfigNodes.SPECIAL_VICTORY_EFFECTS_WEAKNESS_ON_REVOLT_SIEGE_DECISIVE_DEFENDER_VICTORY);
 	}
